Move FileWriter and Rewrite to DataWriter/ folder
