+++ src/DataWriter/Rewrite.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Tekreme73\Laravel\ConfigWriter;
+namespace Tekreme73\Laravel\ConfigWriter\DataWriter;
 
 use Exception;
