+++ tests/Config/RewriteTest.php
@@ -1,7 +1,7 @@
 <?php
 
 use PHPUnit\Framework\TestCase;
-use Tekreme73\Laravel\ConfigWriter\Rewrite;
+use Tekreme73\Laravel\ConfigWriter\DataWriter\Rewrite;
 
 class RewriteTest extends TestCase
 {
